requirements.txt
gradio
rembg[cpu]
Pillow
numpy
